TS-50: Always show non-working project name if not a workday
